Arbiter V3: Chainlink Data Streams integration, modular DEX adapters and major refactoring.

+++ contracts/interfaces/IFlashLiquidityFactory.sol
@@ -3,7 +3,7 @@
 pragma solidity >=0.5.0;
 
 interface IFlashLiquidityFactory {
-    event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+    event PairCreated(address indexed token0, address indexed token1, address pair, uint256);
 
     function feeTo() external view returns (address);
 
@@ -13,9 +13,9 @@ interface IFlashLiquidityFactory {
 
     function getPair(address tokenA, address tokenB) external view returns (address pair);
 
-    function allPairs(uint) external view returns (address pair);
+    function allPairs(uint256) external view returns (address pair);
 
-    function allPairsLength() external view returns (uint);
+    function allPairsLength() external view returns (uint256);
 
     function createPair(address tokenA, address tokenB) external returns (address pair);
